requirements.txt
annotated-types==0.7.0
anyio==4.5.2
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
fsspec==2024.12.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-dialogflow==2.38.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0.dev2
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.28.0
idna==3.10
importlib_resources==6.4.5
jiter==0.8.2
numpy==1.24.4
oauthlib==3.2.2
openai==1.60.2
packaging==24.2
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.1.4
python-dotenv==1.0.1
python-telegram-bot==13.7
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.5.2
six==1.17.0
sniffio==1.3.1
tokenizers==0.20.3
tornado==6.4.2
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
zipp==3.20.2
Flask==2.2.2
python-dotenv==1.0.1
